# Improved System.Decimal performance