--- a/src/System.Private.CoreLib/shared/System/Math.cs
+++ b/src/System.Private.CoreLib/shared/System/Math.cs
@@ -96,7 +96,7 @@ namespace System
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static decimal Abs(decimal value)
         {
-            return decimal.Abs(value);
+            return decimal.Abs(ref value);
         }
 
         [StackTraceHidden]
@@ -453,7 +453,7 @@ namespace System
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static decimal Max(decimal val1, decimal val2)
         {
-            return decimal.Max(val1, val2);
+            return decimal.Max(ref val1, ref val2);
         }
 
         public static double Max(double val1, double val2)
@@ -541,7 +541,7 @@ namespace System
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static decimal Min(decimal val1, decimal val2)
         {
-            return decimal.Min(val1, val2);
+            return decimal.Min(ref val1, ref val2);
         }
 
         public static double Min(double val1, double val2)