@jkotas 2017-11-23
> Is it possible to somehow reference System.Runtime.Intrinsics.X86 in CoreRT? 

Not yet.