@pentp 2018-01-14
This covers the remaining managed functions of CoreRT that use native code in CoreCLR. I added new tests also.
I plan to address the 32bit performance regressions (as compared to native code) separately.