@jkotas 2018-01-15
> Is it possible to somehow reference System.Runtime.Intrinsics.X86 in CoreRT?

Not yet.