@pentp 2018-01-16
@jkotas This should be ready for merging.